requirement.txt
deepface
Pillow
opencv-python==4.3.0.36
pyqt5
pyside2
nvidia-riva-client
numpy==1.22